Backend/requirements.txt
bidict==0.22.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SocketIO==5.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
six==1.16.0
Werkzeug==2.2.2
